backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-socketio==5.11.4
redis[hiredis]==5.2.1
google-generativeai==0.8.4
groq==0.15.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
aiosqlite==0.20.0
httpx==0.28.1
